datafeed/producers/requirements.txt
pandas
requests
asyncio
kafka-python
numpy
datetime
websockets
cryptofeed
aiokafka